requirements.txt
# web框架
Flask==0.10.1

# 在flask中用SQLAlchemy更方便
Flask-SQLAlchemy==2.0

# 模板引擎 
Jinja2==2.7.3

# markdown文本解析
Markdown==2.6.2

# jinja2用的
MarkupSafe==0.23

# 配置文件解析
PyYAML==3.11

# http请求
requests==2.7.0

# 数据库工具
SQLAlchemy==1.0.4

# WSGI函数库
Werkzeug==0.10.4

# xml解析
lxml==3.4.4

# 代码块高亮
Pygments=2.2.0
